refactor: restructure library for accommodating TUI needs
