+++ src/upload/mod.rs
@@ -1,3 +1,5 @@
+//! The upload module provides the package upload functionality.
+
 use futures::TryStreamExt;
 use indicatif::{style::TemplateError, HumanBytes, ProgressState};
 use std::{
@@ -46,6 +48,7 @@ fn get_default_client() -> Result<reqwest::Client, reqwest::Error> {
         .build()
 }
 
+/// Uploads package files to a Quetz server.
 pub async fn upload_package_to_quetz(
     storage: &AuthenticationStorage,
     api_key: Option<String>,
@@ -100,6 +103,7 @@ pub async fn upload_package_to_quetz(
     Ok(())
 }
 
+/// Uploads package files to an Artifactory server.
 pub async fn upload_package_to_artifactory(
     storage: &AuthenticationStorage,
     username: Option<String>,
@@ -165,6 +169,7 @@ pub async fn upload_package_to_artifactory(
     Ok(())
 }
 
+/// Uploads package files to a prefix.dev server.
 pub async fn upload_package_to_prefix(
     storage: &AuthenticationStorage,
     api_key: Option<String>,
@@ -226,6 +231,7 @@ pub async fn upload_package_to_prefix(
     Ok(())
 }
 
+/// Uploads package files to an Anaconda server.
 pub async fn upload_package_to_anaconda(
     storage: &AuthenticationStorage,
     token: Option<String>,
